requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
pysnmp==4.4.12
python-nmap==0.7.1
requests==2.26.0
python-dotenv==0.19.0 